Support for transformating annotation values and source retention policies (#2720)
This change introduces a new interface, `AnnotationTransformer`, that discards the original annotation thus adding the ability for extensions to optimize memory consumption by not retaining metadata that is not used at runtime.

In addition this change provides a solution to transform existing runtime annotations to source level annotations which provides a potential solution for the #2703.

This change is also breaking as it alters Micronaut to properly respect source level retention defined in annotations. Source level annotations will no longer be retained metadata from Micronaut 2.0 onwards.