Register inner classes of classes annotated with @ReflectiveAccess (#5735)
In previous versions of Micronaut the handling of `@Introspected` produces introspections and reflective metadata for inner enums.

Since the changes to remove reflective access from introspected this is no longer the case. This restores the previous behaviour for `@Entity` with inner classes that is causing the Micronaut SQL tests to fail.