Implement correct handling for h2c upgrade request in client and server [Fixes #5005] (#6270)
Before this patch, the h2c server would not respond to the initial request that started the h2c upgrade, and the client would improperly handle the response to its own upgrade request.

This patch changes the server to respond to the first request, and the client to ignore messages on h2c stream 1 (the stream the response arrives on), to make them compatible with the spec. This means that old micronaut h2c clients without this patch are now incompatible with fixed h2c servers.