Mark attribute multipart parts for release (#6511)
Attribute parts were not released properly when combined with streaming another part. This patch adds the attribute part to the request so that the request handles cleanup. We can't release the data immediately, because the route method may still expect it as a ByteBuffer parameter.

There is no unit test for this, because this only fixes a bytebuf leak, which can't be tested in our current setup.

Fixes #4837