Replace usages of identityHashCode to prevent collisions (#6512)
System.identityHashCode is, by default, set to a random int by the JVM. Using it as a map key could lead to collisions. This patch replaces those usages with a new wrapper class that, on top of the identityHashCode for hashing, uses == for the equals check. This prevents any collisions in the map keys.

There is unfortunately no way to test for these collisions without a JVM option (-XX:hashCode), so there is no unit test in this patch.