Fix input file normalization for functional tests (#6558)
* Fix functional tests being re-executed with build cache

This commit fixes functional tests being re-executed after a clean
build when caching is enabled. This was discovered during the Gradle
Enterprise caching experiments, in the following scan:

https://ge.micronaut.io/c/jvle2kuhz75yg/hmskryjn6d2xq/task-inputs

It shows that functional tests are re-executed because some inputs
changed. The inputs turned out to be _classpath_, but they are
in practice compile classpath, so they shouldn't have been executed.

To fix this, we moved the common code to a convention plugin for
functional tests which defines a custom `FunctionalTest` type which
properly declares the classpath as inputs using the `@CompileClasspath`
normalization strategy.

After this fix, re-executing the build after a `clean` doesn't trigger
task execution anymore:

https://ge.micronaut.io/s/h4job5l4syrpc/timeline?hide-timeline&sort=longest&task-path=:test-suite:functionalTest

This will also make the build cacheable _cross machines_ by not
using absolute paths.

It's worth noting that this still requires setting the system properties
at execution time. The PR at https://github.com/gradle/gradle/pull/19037
would make it cleaner because we could configure the task using something
like:

```
systemProperties.put("...", appUnderTestClasspath.map { it.asPath })
```

It would not, however, remove the need for an explicit task type
because this API will not let us declare the normalization strategy.

* Disable parallel/caching

We don't want to enable them for all builds _yet_, only for
experiments.

* Apply the `base` plugin for the BOM checker module

This makes sure that there's a `clean` and `check` lifecycle task.