Update directory watcher and JNA to the first M1 supporting version (#6720)
Running a micronaut app with Gradle in `--continuous` (`-t`) mode doesn't currently
work on M1 Apple hardware.  The old version of directory-watcher relies on JNA 5.3.1
and we have it set to this in the managed versions.

This PR updates the directory-watcher to `io.methvin:directory-watcher:0.15.0` and
updates our managed JNA version similarly so that the transitive dependency is not
downgraded when it is resolved.

We have JNA pulled in transitively by other projects as well:

https://ge.micronaut.io/s/h7uz7ojxqjg4s/dependencies?dependencies=jna&expandAll

These are still on 5.6.0... Will this cause issues when the runtime-osx library is
pulled in to the build?