Allow matching `@Produces(ALL)` for ambiguous routes (#8039)
If the client sends a specific content type, but the controller is annotated `@Produces(ALL)`, the matching logic for ambiguous routes would bail out and match no routes. This patch removes that logic, so that the ambiguity resolution can go on and match e.g. based on path variables.
Fixes #7344