Support type annotations for Java & Groovy (#8764)
Introduced a way to access type annotations on ClassElement#getTypeAnnotationMetadata. To do that the native type has to carry the actual type mirror (for Java). If the type mirror is present that indicates that ClassElement supports type annotations and the actual annotation metadata of the class is a hierarchy of both. Modifying annotation metadata of such class will modify the type annotations, modifying the actual class annotation can be done on the #getType class element.

The type annotations are now the ones that are persisted for an executable method's return type, parameters, and type parameters. There is no need for hacks to persist generic type annotations in some cases.

There is some modification to how the generic placeholders are preserved bound to make sure everything works correctly.

Added a lot of tests to test annotating methods, return types, parameters, fields, and field types.