Fix all-opens support in ksp (#9456)
ksp does not currently see when classes or methods are opened by the all-opens plugin. This patch assumes these elements are open if the annotations that all-opens considers by default are present.

There's no test in this PR, I have one in micronaut-starter. However from the tests that I had to disable in FinalModifierSpec, I'm pretty sure this will work.

Fixes #9426


---------

Co-authored-by: Denis Stepanov <denis.s.stepanov@oracle.com>