Fix NCDFE when snakeyaml is missing (#9467)
I'm a bit confused about when NCDFEs happen and when they don't. From what I can tell, this NCDFE is caused by a Methodref in the constant pool, even though the method that uses the Methodref is never called.

This PR moves the method, and thus the Methodref, to a nested class. There is still a Class entry in the constant pool, but that seems to be fine in testing.

 Fixes #9464