Make CopyOnWriteMap.evict more resilient for ConcurrentHashMap (#9742)
In DefaultMutableConversionService, we use COWMap.evict to evict entries from a CHM concurrently. This could lead to errors if multiple evict operations happened at the same time. This patch handles those errors. It's still possible for too many or too few entries to be evicted, but we can live with that.

Fixes #9741