Implement JsonAnyGetter for BeanIntrospectionModule (#9872)
Normally this would be a feature I skip out on, because we can't support every jackson feature in BeanIntrospectionModule. However in this case, the BIM actually duplicated the AnyGetter as a normal property also, so a fix was necessary anyway, and implementing the full logic is not that difficult and a test case existed already.

A similar problem exists for JsonValue, but because there's no existing test case for it, I don't want to implement it fully. That means JsonValue will still not work with native image, likely. But JsonAnyGetter should.

Fixes #9870