Fix logged error when using form fields with blocking filter (#10021)
The blocking filter would cause onSubscribe to be delayed, leading to a logged exception when trying to request from upstream. This patch simply skips the request call in that case. The request will be done in onSubscribe later.

Also make sure the subscriber methods are only called on the event loop to avoid concurrency issues.