Fix client read timeout between earmark and dispatch (#10050)
Connection.hasLiveRequests was set to true by the earmark operation at the start of Connection.dispatch, but the read timeout is only reset on the event loop just before dispatch0. That means that a read timeout could trigger after hasLiveRequests becomes true but before the reset, leading to an unhandled timeout and wasted connection.

The fix is to separate the "earmark" flag from the hasLiveRequests flag (or counter for HTTP/2). They will usually have the same value, except in the short interval before dispatch0, where earmark is true but hasLiveRequests is not. This means the read timeout is ignored before the timeout is reset.

Fixes #9899