Stacktrace for errors when finalizing processors (#10467)
The TypeElementVisitorProcessor calls the `.finish()` method on visitors and catches all exceptions if present. But the message might not be descriptive enough. For example a StackOverflowError does not have a message, so the logs would be:
`[ERROR] error: Error finalizing type visitor [...Visitor@5e67a490]: null`
Since the visitors are not expected to fail frequently, printing stack trace should not have a negative impact on performance.