Support binding RouteMatch in server filter methods (#10961)
This PR adds support for RouteMatch parameters to server filter methods.

Unfortunately, binding was previously fixed for the server and client. This code had no access to RouteMatch (http does not depend on router). I refactored the code to move the arg binding into BaseFilterProcessor. This is then overridden in the server implementation to add RouteMatch binding support.

I also moved ServerHttpRequest binding there while I was at it.

I've tried to keep the move limited. Most of the arg binding code is still in MethodFilter. This avoids having to expose a dozen private classes.

Fixes #10942