Properly discard inbound HttpContent when connection cancels before readComplete (#11496)
When a connection is closed before readComplete is called, OptimisticBufferingInboundHandler would not release the input buffer correctly.

Discovered by fuzzing with EmbeddedChannel. I'm not sure if this can happen with a non-embedded channel, but better safe than sorry. A test is not feasible, detection relies on finding dangling buffers, which the test suite is too noisy for.