Support Tracing of Dependency Injection (#11449)
Adds the ability to trace the behaviour of dependency and configuration injection. Can be activated a number of different ways, the simplest being an environment variable:

```
MICRONAUT_INJECT_TRACE=.+ ./gradlew run
```

Which enables useful output such as what is the configuration profile of the application, where configuration is read from, what beans are created, how long the bean took to be created, the graph of the bean creation and where it was created from.

To support this, the PR also adds property origin tracking so that one can inspect where a property originated from. 

In addition, some of the changes in `DefaultBeanContext` are in order to reduce stack trace sizes which were starting to get out of hand.