Fix duplicate error in HTTP client (#11727)
When there's two errors, the DefaultHttpClient would try to complete the same future twice, leading to a logged message. This change adds a state transition so that the second exception is handled by AfterContent, which leads to pool-level handling instead.

A test is not possible because the only thing fixed here is a superfluous log statement.