Spike `pyronaut` command line
This commit introduces a `pyronaut` command-line which allows running
Pyronaut application without the user seeing that this is actually a
Java application. For now the CLI cannot be used as a native image,
but it does support live reloading.

To try it:

- run `./gradlew installPyronautDist`
- export PATH=$PATH:<path to micronaut-core>/pyronaut-cli/build/install/micronaut-pyronaut-cli-pyronaut/bin
- `cd /tmp`
- `mkdir -p demo/foo`
- `cd demo`

Create a test controller in the `foo` directory (Python scripts at the
root won't work for some reason for now):

```
from micronaut.http.annotation import Controller, Get

@Controller
class MyController:
    @Get(value="/", produces="text/plain")
    def index(self) -> str:
        return "Hello World!"

    @Get(value="/hello")
    def hello(self) -> dict:
        return {"Hello": "Pyronaut!"}
```

Then run the app:

`pyronaut run`